Trait reth_prune::segments::Segment

pub trait Segment<DB: Database>: Debug + Send + Sync {
    // Required methods
    fn segment(&self) -> PruneSegment;
    fn mode(&self) -> Option<PruneMode>;
    fn prune(
        &self,
        provider: &DatabaseProviderRW<DB>,
        input: PruneInput
    ) -> Result<PruneOutput, PrunerError>;

    // Provided method
    fn save_checkpoint(
        &self,
        provider: &DatabaseProviderRW<DB>,
        checkpoint: PruneCheckpoint
    ) -> ProviderResult<()> { ... }
}

A segment represents a pruning of some portion of the data.

Segments are called from Pruner with the following lifecycle:

Call Segment::prune with delete_limit of PruneInput.

If Segment::prune returned a Some in checkpoint of PruneOutput, call Segment::save_checkpoint.

Subtract pruned of PruneOutput from delete_limit of next PruneInput.

Required Methods

fn segment(&self) -> PruneSegment

Segment of data that’s pruned.

fn mode(&self) -> Option<PruneMode>

Prune mode with which the segment was initialized

fn prune( &self, provider: &DatabaseProviderRW<DB>, input: PruneInput ) -> Result<PruneOutput, PrunerError>

Prune data for Self::segment using the provided input.

Provided Methods

fn save_checkpoint( &self, provider: &DatabaseProviderRW<DB>, checkpoint: PruneCheckpoint ) -> ProviderResult<()>

Save checkpoint for Self::segment to the database.